Updated react, axios, npm Added jest Extract Service offerings from app
