+++ package.json
@@ -6,7 +6,7 @@
     "@axios-use/react": "^6.0.0",
     "@reduxjs/toolkit": "^1.8.3",
     "@szhsin/react-menu": "^3.1.1",
-    "axios": "^0.26.0",
+    "axios": "^1.4.0",
     "history": "^5.3.0",
     "i18next": "^21.6.12",
     "i18next-browser-languagedetector": "^6.1.3",
@@ -17,23 +17,23 @@
     "mdbreact": "^5.2.0",
     "mem": "^9.0.2",
     "moment": "^2.29.3",
-    "npm": "^8.10.0",
+    "npm": "^9.8.1",
     "plotly.js": "^2.12.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-collapse": "^5.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-file-download": "^0.3.5",
     "react-i18next": "^11.15.5",
-    "react-image-gallery": "^1.2.8",
-    "react-multi-carousel": "^2.8.2",
-    "react-paginate": "^8.1.3",
-    "react-plotly.js": "^2.5.1",
+    "react-image-gallery": "^1.3.0",
+    "react-multi-carousel": "^2.8.4",
+    "react-paginate": "^8.2.0",
+    "react-plotly.js": "^2.6.0",
     "react-redux": "^7.2.8",
     "react-responsive-carousel": "^3.2.23",
-    "react-responsive-modal": "^6.2.0",
+    "react-responsive-modal": "^6.4.2",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "react-tooltip": "^4.2.21",
     "redux-thunk": "^2.4.1",
     "styled-components": "^5.3.5",
@@ -65,11 +65,14 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.0.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "customize-cra": "^1.0.0",
     "eslint-plugin-react": "^7.29.4",
+    "jest": "^29.6.2",
+    "jest-junit": "^16.0.0",
     "react-app-rewired": "^2.2.1",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0"
+    "ts-jest": "^29.1.1"
   }
 }
