+++ src/components/service-offerings/ServiceOfferings.css
@@ -1,0 +1,12 @@
+header {
+    background: linear-gradient(110deg, rgb(185, 0, 255), 15%, rgb(0, 0, 148), 75%, rgb(70, 218, 255));
+    height:100px;
+}
+
+h2 {
+    color: white;
+    padding: 30px 0 0 20px;
+}
+.content {
+    margin-top:20px;
+}
