+++ src/index.js
@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from 'react-redux';
 import './index.css';
@@ -10,7 +10,6 @@ import axios from "axios"
 import { retrieveToken } from "./common/auth";
 import { Navigate, useNavigate } from 'react-router-dom';
 import history from "./common/history"
-
 
 // axios.interceptors.request.use(function (config) {
 //   const token = retrieveToken();
@@ -67,12 +66,13 @@ store.subscribe(() => {
   saveToLocalStorage(store.getState());
 });
 
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
 
-ReactDOM.render(
+root.render(
   <Provider store={store}>
     <Suspense fallback="..... is loading" >
       <App />
     </Suspense>
   </Provider>,
-  document.getElementById('root')
 );
